feat(config): add glob matching logic
